+++ packages/gasket-cli/test/unit/scaffold/actions/setup-pkg.test.js
@@ -17,10 +17,10 @@ describe('setupPkg', () => {
     mockContext = {
       appName: 'my-app',
       appDescription: 'my cool app',
-      rawPresets: ['@gasket/bogus-preset'],
-      presetPkgs: [{
+      presetInfos: [{
         name: '@gasket/bogus-preset',
-        version: '3.2.1'
+        version: '3.2.1',
+        rawName: '@gasket/bogus-preset'
       }],
       cliVersionRequired: '^1.2.3'
     };
@@ -35,48 +35,48 @@ describe('setupPkg', () => {
   });
 
   it('instantiates PackageJson app name and description', async () => {
-    await setupPkg(mockContext);
+    await setupPkg.wrapped(mockContext);
     assume(packageJsonSpy).is.called();
     assume(packageJsonSpy.args[0][0]).property('name', mockContext.appName);
     assume(packageJsonSpy.args[0][0]).property('description', mockContext.appDescription);
   });
 
   it('adds the preset to pkg dependencies', async () => {
-    await setupPkg(mockContext);
+    await setupPkg.wrapped(mockContext);
     assume(mockContext.pkg.fields).property('dependencies');
     assume(mockContext.pkg.fields.dependencies).property('@gasket/bogus-preset');
   });
 
   it('adds the preset version as compatible with package', async () => {
-    await setupPkg(mockContext);
+    await setupPkg.wrapped(mockContext);
     assume(mockContext.pkg.fields).property('dependencies');
     assume(mockContext.pkg.fields.dependencies).property('@gasket/bogus-preset', '^3.2.1');
   });
 
   it('adds the preset version from rawPresets', async () => {
-    mockContext.rawPresets = ['@gasket/bogus-preset@4.5.6'];
-    await setupPkg(mockContext);
+    mockContext.presetInfos[0].rawName = '@gasket/bogus-preset@4.5.6';
+    await setupPkg.wrapped(mockContext);
     assume(mockContext.pkg.fields).property('dependencies');
     assume(mockContext.pkg.fields.dependencies).property('@gasket/bogus-preset', '4.5.6');
   });
 
   it('adds the preset version with presetPath', async () => {
-    mockContext.rawPresets = ['@gasket/bogus-preset@file:/path/to/preset'];
-    await setupPkg(mockContext);
+    mockContext.presetInfos[0].rawName = '@gasket/bogus-preset@file:/path/to/preset';
+    await setupPkg.wrapped(mockContext);
     assume(mockContext.pkg.fields).property('dependencies');
     assume(mockContext.pkg.fields.dependencies).property('@gasket/bogus-preset', 'file:/path/to/preset');
   });
 
   it('adds extra plugins to pkg dependencies', async () => {
     mockContext.rawPlugins = ['jest@1.2.3', 'my-custom-gasket-plugin'];
-    await setupPkg(mockContext);
+    await setupPkg.wrapped(mockContext);
     assume(mockContext.pkg.fields.dependencies).property('@gasket/jest-plugin', '1.2.3');
     assume(mockContext.pkg.fields.dependencies).property('my-custom-gasket-plugin', 'latest');
   });
 
   it('adds pkg to context', async () => {
     assume(mockContext.pkg).is.undefined();
-    await setupPkg(mockContext);
+    await setupPkg.wrapped(mockContext);
     assume(mockContext.pkg).is.instanceOf(ConfigBuilder);
   });
 });