+++ packages/gasket-cli/CHANGELOG.md
@@ -1,5 +1,8 @@
 # `@gasket/cli`
 
+- Use Loader for presets during create ([#66])
+  - Presets can export createContext from module
+  - `PresetInfo`s for loaded modules available on createContext
 - Enable metadata for create hooks ([#64])
 
 ### 2.6.1
@@ -213,3 +216,4 @@ upon `gasket create`
 [#38]: https://github.com/godaddy/gasket/pull/38
 [#58]: https://github.com/godaddy/gasket/pull/58
 [#64]: https://github.com/godaddy/gasket/pull/64
+[#66]: https://github.com/godaddy/gasket/pull/66